Rename ConnectionQueue to JoinQueue
