+++ Content.Client/Corvax/JoinQueue/QueueGui.xaml
