+++ Content.Client/Corvax/JoinQueue/QueueGui.xaml.cs
@@ -4,7 +4,7 @@ using Robust.Client.UserInterface;
 using Robust.Client.UserInterface.Controls;
 using Robust.Client.UserInterface.XAML;
 
-namespace Content.Client.Corvax.ConnectionQueue;
+namespace Content.Client.Corvax.JoinQueue;
 
 [GenerateTypedNameReferences]
 public sealed partial class QueueGui : Control
