+++ Content.Client/Entry/EntryPoint.cs
@@ -1,7 +1,7 @@
 using Content.Client.Administration.Managers;
 using Content.Client.Changelog;
 using Content.Client.Chat.Managers;
-using Content.Client.Corvax.ConnectionQueue;
+using Content.Client.Corvax.JoinQueue;
 using Content.Client.Corvax.Sponsors;
 using Content.Client.Options;
 using Content.Client.Eui;
@@ -75,7 +75,7 @@ namespace Content.Client.Entry
         [Dependency] private readonly PlayTimeTrackingManager _playTimeTracking = default!;
         [Dependency] private readonly ContentLocalizationManager _contentLoc = default!;
         [Dependency] private readonly ClientSponsorsManager _sponsorsManager = default!; // Corvax-Sponsors
-        [Dependency] private readonly ConnectionQueueManager _queueManager = default!; // Corvax-Queue
+        [Dependency] private readonly JoinQueueManager _queueManager = default!; // Corvax-Queue
 
         public const int NetBufferSizeOverride = 2;
