+++ Content.Client/IoC/ClientContentIoC.cs
@@ -2,7 +2,7 @@
 using Content.Client.Changelog;
 using Content.Client.Chat.Managers;
 using Content.Client.Clickable;
-using Content.Client.Corvax.ConnectionQueue;
+using Content.Client.Corvax.JoinQueue;
 using Content.Client.Corvax.Sponsors;
 using Content.Client.Options;
 using Content.Client.Eui;
@@ -45,7 +45,7 @@ namespace Content.Client.IoC
             IoCManager.Register<ExtendedDisconnectInformationManager>();
             IoCManager.Register<PlayTimeTrackingManager>();
             IoCManager.Register<ClientSponsorsManager>(); // Corvax-Sponsors
-            IoCManager.Register<ConnectionQueueManager>(); // Corvax-Queue
+            IoCManager.Register<JoinQueueManager>(); // Corvax-Queue
         }
     }
 }
