+++ Content.Server/Entry/EntryPoint.cs
@@ -5,7 +5,7 @@ using Content.Server.Administration.Managers;
 using Content.Server.Afk;
 using Content.Server.Chat.Managers;
 using Content.Server.Connection;
-using Content.Server.Corvax.ConnectionQueue;
+using Content.Server.Corvax.JoinQueue;
 using Content.Server.Corvax.Sponsors;
 using Content.Server.Database;
 using Content.Server.EUI;
@@ -99,7 +99,7 @@ namespace Content.Server.Entry
                 IoCManager.Resolve<NetworkResourceManager>().Initialize();
                 IoCManager.Resolve<GhostKickManager>().Initialize();
                 IoCManager.Resolve<ServerSponsorsManager>().Initialize(); // Corvax-Sponsors
-                IoCManager.Resolve<ConnectionQueueManager>().Initialize(); // Corvax-Sponsors
+                IoCManager.Resolve<JoinQueueManager>().Initialize(); // Corvax-Sponsors
 
                 _voteManager.Initialize();
                 _updateManager.Initialize();
