+++ Content.Server/GameTicking/GameTicker.StatusShell.cs
@@ -1,5 +1,5 @@
 using System.Text.Json.Nodes;
-using Content.Server.Corvax.ConnectionQueue;
+using Content.Server.Corvax.JoinQueue;
 using Content.Shared.CCVar;
 using Robust.Server.ServerStatus;
 using Robust.Shared.Configuration;
@@ -23,7 +23,7 @@ namespace Content.Server.GameTicking
         ///     For access to CVars in status responses.
         /// </summary>
         [Dependency] private readonly IConfigurationManager _cfg = default!;
-        [Dependency] private readonly ConnectionQueueManager _queueManager = default!; // Corvax-Queue
+        [Dependency] private readonly JoinQueueManager _queueManager = default!; // Corvax-Queue
 
         private void InitializeStatusShell()
         {
